symbol of FRED economic data

ed_fred_symbol provides an interface to search symbols of economic data from FRED by category or keywords.

ed_fred_symbol(category = NULL, keywords = NULL, ...)

Arguments

category
the category id. If it is NULL, then search symbols from the top categories step by step.

keywords
the query text. If it is NULL, the function will search symbols by category.

...
ignored parameters

Examples

# \donttest{
# search symbols by category
# from top categories
symbol_dt1 = ed_fred_symbol()

#>       id                                    name parent_id
#> 1: 32991 Money, Banking, & Finance                       0
#> 2:    10 Population, Employment, & Labor Markets         0
#> 3: 32992 National Accounts                               0
#> 4:     1 Production & Business Activity                  0
#> 5: 32455 Prices                                          0
#> 6: 32263 International Data                              0
#> 7:  3008 U.S. Regional Data                              0
#> 8: 33060 Academic Data                                   0
#> select one row via ('r'+rowid) or (id): 
#>       id                          name parent_id
#> 1:    22 Interest Rates                    32991
#> 2:    15 Exchange Rates                    32991
#> 3:    24 Monetary Data                     32991
#> 4:    46 Financial Indicators              32991
#> 5:    23 Banking                           32991
#> 6: 32360 Business Lending                  32991
#> 7: 32145 Foreign Exchange Intervention     32991
#> select one row via ('r'+rowid) or (id): 
#>        id                                            name parent_id
#>  1: 33058 Automobile Loan Rates                                  22
#>  2:    51 Bankers Acceptance Rate                                22
#>  3:   121 Certificates of Deposit                                22
#>  4:   120 Commercial Paper                                       22
#>  5: 32348 Corporate Bonds                                        22
#>  6: 33059 Credit Card Loan Rates                                 22
#>  7: 32298 Eurodollar Deposits                                    22
#>  8:   118 FRB Rates - discount, fed funds, primary credit        22
#>  9: 33056 Interest Checking Accounts                             22
#> 10: 33446 Interest Rate Spreads                                  22
#> 11: 32299 Interest Rate Swaps                                    22
#> 12: 33003 LIBOR Rates                                            22
#> 13: 32995 Long-Term Securities                                   22
#> 14: 33055 Money Market Accounts                                  22
#> 15:   114 Mortgage Rates                                         22
#> 16: 33057 Personal Loan Rates                                    22
#> 17:   117 Prime Bank Loan Rate                                   22
#> 18: 33491 Saving Accounts                                        22
#> 19:   116 Treasury Bills                                         22
#> 20:   115 Treasury Constant Maturity                             22
#> 21:    82 Treasury Inflation-Indexed Securities                  22
#>        id                                            name parent_id
#> select one row via ('r'+rowid) or (id): 
#>             symbol
#> 1: TERMCBAUTO48NS 
#> 2: RIFLPBCIANM60NM
#> 3: TERMAFCNCNSA   
#>                                                                                       name
#> 1: Finance Rate on Consumer Installment Loans at Commercial Banks, New Autos 48 Month Loan
#> 2: Finance Rate on Consumer Installment Loans at Commercial Banks, New Autos 60 Month Loan
#> 3: New Car Average Finance Rate at Auto Finance Companies (DISCONTINUED)                  
#>    frequency   units
#> 1:   Monthly Percent
#> 2:   Monthly Percent
#> 3:   Monthly Percent
#> select rows via ('r'+rowid) or (symbol): 

# specify the initial categories
symbol_dt2 = ed_fred_symbol(category = 1)

#>        id                                         name parent_id
#>  1:     1 Parent Series                                        1
#>  2: 32262 Business Cycle Expansions & Contractions             1
#>  3: 32436 Construction                                         1
#>  4: 33490 Finance Companies                                    1
#>  5: 32216 Health Insurance                                     1
#>  6:    97 Housing                                              1
#>  7:     3 Industrial Production & Capacity Utilization         1
#>  8: 32429 Manufacturing                                        1
#>  9:     6 Retail Trade                                         1
#> 10: 33441 Services                                             1
#> 11: 33492 Technology                                           1
#> 12: 33202 Transportation                                       1
#> 13: 33203 Wholesale Trade                                      1
#> select one row via ('r'+rowid) or (id): 

#> Error in menu(choices, cat(sprintf("Select the value of %s", arg_name))): menu() cannot be used non-interactively

# search symbol by keywords
symbol_dt3 = ed_fred_symbol(keywords = "gdp china")

#> Error in menu(choices, cat(sprintf("Select the value of %s", arg_name))): menu() cannot be used non-interactively

# }