creating weighted index

pq_index creates a sector/industry index using the method of weighted geometric mean, based on a set of data and corresponding weights.

pq_index(dt, w = "cap_total", x = "close|value", base_value = 1,
  base_date = NULL, name = NULL)

Arguments

dt
a list/dataframe of time series dataset

w
the name of weights column. Default is 'cap_total'. If x is not available or is NULL, then using equal weights.

x
the name of column to create index. Default is 'close|value'

base_value
the base value of index. Default is 1.

base_date
the base date of index. Default is the minimum date.

name
the name of index. Default is NULL, then using 'index'.

Examples

# \donttest{
# Example I bank share index
library(data.table)
data(ssec)
bank_dat = setDT(ssec)[symbol != '000001.SS']
# creating index
bank_index = pq_index(bank_dat, x='close', w='cap_total')
# pq_plot(bank_index)

# }