calculating performance metrics

pq_performance calculates performance metrics based on returns of market price or portfolio. The performance analysis functions are calling from PerformanceAnalytics package, which includes many widely used performance metrics.

pq_performance(dt, Ra, Rb = NULL, perf_fun, ...)

Arguments

dt
a list/dataframe of time series datasets.

Ra
the column name of asset returns.

Rb
the column name of baseline returns, defaults to NULL.

perf_fun
performance function from PerformanceAnalytics package, see pq_perf_funs.

...
additional parameters, the arguments used in PerformanceAnalytics functions.

Examples

 
# \donttest{
library(data.table)
library(pedquant) 

# load data
data(ssec)
# transfer prices to returns
datret = pq_return(ssec, 'close', freq = 'monthly', col_rename = 'Ra')

#> 1/6 000001.SS
#> 2/6 601288.SS
#> 3/6 601328.SS
#> 4/6 601398.SS
#> 5/6 601939.SS
#> 6/6 601988.SS

datret2 = rbindlist(datret)[, date := as.Date(sub('[0-9]{2}$', '01', date))][]

# merge returns of assets and baseline
datRaRb = merge(
   datret2[symbol != '000001.SS', .(date, symbol, Ra)], 
    datret2[symbol == '000001.SS', .(date, Rb = Ra)],
    by = 'date', all.x = TRUE
)

# claculate table.CAPM metrics
perf_capm = pq_performance(datRaRb, Ra = 'Ra', Rb = 'Rb', perf_fun = 'table.CAPM')
rbindlist(perf_capm, idcol = 'symbol')

#>       symbol   Alpha   Beta  Beta+  Beta- R-squared Annualized Alpha
#> 1: 601988.SS -0.0025 0.7586 0.9856 0.7915    0.5908          -0.0295
#> 2: 601398.SS  0.0014 0.7656 1.0752 0.6446    0.5076           0.0175
#> 3: 601328.SS -0.0026 0.9254 1.0410 0.8834    0.6105          -0.0302
#> 4: 601939.SS  0.0030 0.8601 1.0028 0.8156    0.5822           0.0364
#> 5: 601288.SS  0.0018 0.5965 0.7900 0.5427    0.4692           0.0223
#>    Correlation Correlation p-value Tracking Error Active Premium
#> 1:      0.7686                   0         0.1974        -0.0476
#> 2:      0.7125                   0         0.2317        -0.0041
#> 3:      0.7814                   0         0.2074        -0.0450
#> 4:      0.7630                   0         0.2024         0.0293
#> 5:      0.6850                   0         0.1680         0.0085
#>    Information Ratio Treynor Ratio
#> 1:           -0.2409       -0.0053
#> 2:           -0.0178        0.0528
#> 3:           -0.2170       -0.0702
#> 4:            0.1449       -0.0195
#> 5:            0.0505        0.0591

# }