creating the equity trends for portfolio

pq_portfolio creates the equity trends of a portfolio secturites.

pq_portfolio(dt, w, x = "close", init_equity = NULL, date_range = "max",
  from = NULL, to = Sys.Date())

Arguments

dt
a list/dataframe of time series dataset

w
a dataframe of position and price of all transactions.

x
the name of column to calculate. Default is 'close|value'.

init_equity
initial equity

date_range
date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is max.

from
the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.

to
the end date. Default is the current date.

Examples

# \donttest{
dat = md_stock(c('FB', 'AMZN', 'AAPL', 'NFLX', 'GOOG'), date_range = 'max')

#> 1/5 FB
#> 2/5 AMZN
#> 3/5 AAPL
#> 4/5 NFLX
#> 5/5 GOOG

w = data.frame(
      date = c('2016-01-08'),
      symbol = c('FB', 'AMZN', 'AAPL', 'NFLX', 'GOOG'),
      price = c(99.88, 618.88, 97.88, 115.88, 716.88),
      position = c(100, 200, 300, 400, 500),
      type = rep('buy_op', 5)
    )
pf = pq_portfolio(dat, w)
# }