calculating returns/equity of portfolio

pq_portfolio calculates the weighted returns or the equity of a portfolio assets.

pq_portfolio(dt, dtv, x, v = "volume", init_fund = NULL,
  method = "arithmetic", ...)

Arguments

dt

a list/dataframe of price by asset.

dtv

a dataframe of transaction volume by asset.

x

the column name of adjusted asset price.

v

the column name of asset volume, defaults to volume.

init_fund

initial fund value.

method

the method to calculate asset returns, the available values include arithmetic and log, defaults to arithmetic.

...

ignored

Examples

library(pedquant)
library(data.table)

data(dt_banks)
datadj = md_stock_adjust(dt_banks, adjust = FALSE)

# example I 
dtv = data.table(
    symbol = c("601288.SS","601328.SS","601398.SS","601939.SS","601988.SS"), 
    volume = c(100, 200, 300, 300, 100)
)
dtRa = pq_portfolio(datadj, x='close_adj', dtv=dtv) 
#> Warning: The initial value is setting to 5569.29
e1 = pq_plot(dtRa, y = 'cumreturns')
e1[[1]]

dtRb = pq_return(dt_ssec, x = 'close', freq = 'daily', cumreturns = TRUE)
e2 = pq_plot(list(Ra = dtRa, Rb = dtRb$`000001.SS`), y = 'cumreturns', 
        arrange = list(rows=1, cols=1))
e2[[1]]

# example II 
dtv = data.table(
    symbol = rep(c("601288.SS","601328.SS","601398.SS","601939.SS","601988.SS"), 3), 
    date = rep(c('2009-03-02', '2010-01-04', '2014-09-01'), each = 5), 
    volume = rep(c(100, 200, 300, 300, 100), 3) * rep(c(1, -1, 2), each = 5)
)
dtRa2 = pq_portfolio(datadj, x='close_adj', dtv=dtv, init_fund = 10000) 
e3 = pq_plot(dtRa2, y = 'balance', 
        addti = list(equity = list(), fund = list()))
e3[[1]]