calculating returns by frequency

pq_return calculates returns for daily series based on specified column, frequency and method type.

pq_return(dt, x, freq = "monthly", method = "arithmetic",
  rcol_name = "return", cols_keep = NULL, date_range = "max",
  from = NULL, to = Sys.Date(), print_step = 1L)

Arguments

dt
a list/dataframe of daily series dataset

x
the column name of adjusted asset price.

freq
the frequency of returns. It supports c('all', 'daily', 'weekly', 'monthly', 'quarterly', 'yearly').

method
the method to calculate returns.

rcol_name
setting the name of return column, defaults to return.

cols_keep
the columns keep in the return data. The columns of symbol, name and date will always kept if they are exist in the input data.

date_range
date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is max.

from
the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.

to
the end date. Default is the current date.

print_step
a non-negative integer. Print symbol name by each print_step iteration. Default is 1L.

Examples

# \donttest{
data(ssec)

# create a close_adj column
datadj = md_stock_adjust(ssec, adjust = FALSE)

# set freq
dts_returns1 = pq_return(datadj, x = 'close_adj', freq = 'all')
#> 1/6 000001.SS
#> 2/6 601288.SS
#> 3/6 601328.SS
#> 4/6 601398.SS
#> 5/6 601939.SS
#> 6/6 601988.SS

# set method
dts_returns2 = pq_return(datadj, x = 'close_adj', method = 'log')
#> 1/6 000001.SS
#> 2/6 601288.SS
#> 3/6 601328.SS
#> 4/6 601398.SS
#> 5/6 601939.SS
#> 6/6 601988.SS

# set cols_keep
dts_returns3 = pq_return(datadj, x = 'close_adj', cols_keep = 'cap_total')
#> 1/6 000001.SS
#> 2/6 601288.SS
#> 3/6 601328.SS
#> 4/6 601398.SS
#> 5/6 601939.SS
#> 6/6 601988.SS

# }